ENH: Add randomly generated sessions plan
This PR includes a script to generate a sessions plan for acquisition,
setting a seed for reproducibility.

A sessions plan has been generated, by setting the seed at the onset
date of the study (2023-10-20).

The plan is rendered twice in the SOPs, first at protocol selection time
and the second before you start acquiring.

Resolves: #38.