enh: alternative implementation to events files with Pandas
Reusing some of the foundations from #272, this PR adds the code to
generate events files from *Psychopy*.

Alongside the PR, the data repository has been updated with the QCT and
BHT events files that could directly be generated from the logs.

This PR should be revised to generate all dwi events (fixation) and rest
(movie), which often fail.

The PR adds a jupyter notebook in which these can be handled with more
granularity.